% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pylintR.R
\name{pylint}
\alias{pylint}
\title{Run 'pylint' on a file or a folder}
\usage{
pylint(modules)
}
\arguments{
\item{modules}{one or more Python files, or a folder containing Python files}
}
\value{
A \code{htmlwidget} object.
}
\description{
Run 'pylint' on a file or a folder.
}
