% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pylintR.R
\name{pylintR-shiny}
\alias{pylintR-shiny}
\alias{pylintROutput}
\alias{renderPylintR}
\title{Shiny bindings for pylintR}
\usage{
pylintROutput(outputId, width = "100\%", height = "400px")

renderPylintR(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended}

\item{expr}{an expression that generates an output of \code{\link{pylint}}}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{whether \code{expr} is a quoted expression
(with \code{quote()}); this is useful if you want to save an expression
in a variable}
}
\value{
\code{pylintROutput} returns an output element that can be included
  in a Shiny UI definition, and \code{renderPylintR} returns a
  \code{shiny.render.function} object that can be included in a Shiny
  server definition.
}
\description{
Output and render functions for using 'pylintR' within Shiny
applications and interactive Rmd documents.
}
